chore(refactor): move isReadonly and checkSharePermissions to FileService
